requirements.txt
fastapi
uvicorn
httpx
python-dotenv
beautifulsoup4
playwright
pandas
numpy
matplotlib
networkx
requests
pdfplumber
PyPDF2
tabula-py
